# webAdmin/openapi.yaml
openapi: 3.1.0
info:
  title: Roulette Reward Backend API
  version: 1.1.0
  description: |
    실제 컨트롤러 구현 기준 API 명세입니다.
    인증은 JWT가 아닌 커스텀 헤더(`X-USER-ID`, `X-ADMIN-TOKEN`)를 사용합니다.
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/v1/products:
    get:
      tags: [Product]
      operationId: getProducts
      summary: 상품 목록 조회
      parameters:
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/ActiveOnlyQuery'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseProductList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/orders:
    post:
      tags: [Order]
      operationId: createOrder
      summary: 상품 주문
      parameters:
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: 주문 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrder'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/orders/me:
    get:
      tags: [Order]
      operationId: getMyOrders
      summary: 내 주문 목록 조회
      parameters:
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/OrderStatusQuery'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/budgets/{date}:
    get:
      tags: [AdminBudget]
      operationId: getDailyBudgetByDate
      summary: 일일 예산 조회(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
        - $ref: '#/components/parameters/DatePath'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseBudget'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [AdminBudget]
      operationId: updateDailyBudgetByDate
      summary: 일일 예산 수정(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
        - $ref: '#/components/parameters/DatePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDailyBudgetRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseBudget'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/admin/products:
    get:
      tags: [AdminProduct]
      operationId: getAdminProducts
      summary: 상품 목록 조회(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseProductList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [AdminProduct]
      operationId: createProduct
      summary: 상품 생성(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '200':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseProduct'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/products/{productId}:
    patch:
      tags: [AdminProduct]
      operationId: updateProduct
      summary: 상품 수정(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
        - $ref: '#/components/parameters/ProductIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseProduct'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [AdminProduct]
      operationId: deleteProduct
      summary: 상품 삭제(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
        - $ref: '#/components/parameters/ProductIdPath'
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUnit'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/orders/{orderId}/cancel:
    post:
      tags: [AdminOrder]
      operationId: cancelOrder
      summary: 주문 취소(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
        - $ref: '#/components/parameters/OrderIdPath'
      responses:
        '200':
          description: 취소 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCancelOrder'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/orders:
    get:
      tags: [AdminOrder]
      operationId: getAdminOrders
      summary: 주문 목록 조회(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
        - $ref: '#/components/parameters/OrderStatusQuery'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/orders/{orderId}/status:
    patch:
      tags: [AdminOrder]
      operationId: updateOrderStatus
      summary: 주문 상태 변경(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
        - $ref: '#/components/parameters/OrderIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderListItem'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    XUserIdHeader:
      name: X-USER-ID
      in: header
      required: true
      description: 사용자 식별자(1 이상의 정수)
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 101
    XAdminTokenHeader:
      name: X-ADMIN-TOKEN
      in: header
      required: true
      description: 어드민 인증 토큰
      schema:
        type: string
      example: ADMIN-STATIC-TOKEN
    DatePath:
      name: date
      in: path
      required: true
      schema:
        type: string
        format: date
      example: '2026-02-15'
...
    ProductIdPath:
      name: productId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1001
    OrderIdPath:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 7001
...
    ActiveOnlyQuery:
      name: activeOnly
      in: query
      required: false
      schema:
        type: boolean
        default: true
      example: true
    OrderStatusQuery:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [PLACED, COMPLETED, CANCELED]
      example: PLACED
  responses:
    BadRequestError:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    UnauthorizedError:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ForbiddenError:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFoundError:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ConflictError:
      description: 상태 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    InternalServerError:
      description: 내부 서버 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, message, details, trace_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        trace_id:
          type: string
    ErrorDetail:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
...
    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    OrderStatus:
      type: string
      enum: [PLACED, COMPLETED, CANCELED]
...
    BudgetResponse:
      type: object
      required: [date, totalBudget, usedBudget, remainingBudget]
      properties:
        date:
          type: string
          format: date
        totalBudget:
          type: integer
        usedBudget:
          type: integer
        remainingBudget:
          type: integer
...
    ProductResponse:
      type: object
      required: [productId, name, description, pricePoints, stock, status, createdAt, updatedAt]
      properties:
        productId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        pricePoints:
          type: integer
        stock:
          type: integer
        status:
          $ref: '#/components/schemas/ProductStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductListResponse:
      type: object
      required: [totalItems, items]
      properties:
        totalItems:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
    CreateOrderRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
          format: int64
          minimum: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 20
    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
    OrderResponse:
      type: object
      required: [orderId, userId, productId, quantity, usedPoints, status, orderedAt]
      properties:
        orderId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
        usedPoints:
          type: integer
        status:
          $ref: '#/components/schemas/OrderStatus'
        orderedAt:
          type: string
          format: date-time
    OrderListItemResponse:
      type: object
      required: [orderId, userId, productId, quantity, usedPoints, status, orderedAt]
      properties:
        orderId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
        usedPoints:
          type: integer
        status:
          $ref: '#/components/schemas/OrderStatus'
        orderedAt:
          type: string
          format: date-time
        canceledAt:
          type: string
          format: date-time
          nullable: true
    OrderListResponse:
      type: object
      required: [totalItems, items]
      properties:
        totalItems:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderListItemResponse'
    UpdateDailyBudgetRequest:
      type: object
      required: [totalBudget]
      properties:
        totalBudget:
          type: integer
          minimum: 0
...
    CreateProductRequest:
      type: object
      required: [name, description, pricePoints, stock, status]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 500
        pricePoints:
          type: integer
          minimum: 1
          maximum: 1000000
        stock:
          type: integer
          minimum: 0
          maximum: 1000000
        status:
          $ref: '#/components/schemas/ProductStatus'
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 500
        pricePoints:
          type: integer
          minimum: 1
          maximum: 1000000
        stock:
          type: integer
          minimum: 0
          maximum: 1000000
        status:
          $ref: '#/components/schemas/ProductStatus'
    CancelOrderResponse:
      type: object
      required: [orderId, userId, refundedPoints, refundedAt]
      properties:
        orderId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        refundedPoints:
          type: integer
        refundedAt:
          type: string
          format: date-time
...
    ApiResponseBudget:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/BudgetResponse'
...
    ApiResponseProductList:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/ProductListResponse'
    ApiResponseOrder:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/OrderResponse'
    ApiResponseOrderList:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/OrderListResponse'
    ApiResponseOrderListItem:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/OrderListItemResponse'
...
    ApiResponseProduct:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/ProductResponse'
    ApiResponseCancelOrder:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/CancelOrderResponse'
    ApiResponseUnit:
      type: object
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: false